refactor(core): own git vcs implementation in plugin
